WIN_requirements.txt
djangorestframework
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
cffi==1.12.3
confusable-homoglyphs==3.2.0
Django==2.2.1
django-bootstrap4==0.0.8
django-celery-beat==1.4.0
django-filter==2.1.0
django-registration==3.0
django-registration-redux==2.5
django-smtp-ssl==1.0
django-timezone-field==3.0
kombu==4.6.3
mysqlclient==1.4.2.post1
Pillow==6.0.0
pycparser==2.19
PyNaCl==1.3.0
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.1
sqlparse==0.3.0
vine==1.3.0